Fix owner ID unpacking bug
The entire u64 can be non-zero, but the lower 16 bits are zero.
With this plus the log size increase, all benchmarks complete
successfully with debug + validation.

+++ cxlalloc/src/slab/shared.rs
@@ -51,7 +51,7 @@ impl Owner {
     }
 
     pub(crate) fn id(&self) -> Option<thread::Id> {
-        Option::<_>::unpack(self.0)
+        Option::<_>::unpack((self.0) as u16 as u64)
     }
 }
